# api/docs/swagger.yaml
definitions:
...
  models.Error:
    properties:
      message:
        type: string
    type: object
  models.LanguageCreate:
    properties:
      name:
        type: string
      picture:
        type: string
    type: object
  models.LanguageResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      picture:
        type: string
      updated_at:
        type: string
    type: object
  models.LanguageUpdate:
    properties:
      id:
        type: integer
      name:
        type: string
      picture:
        type: string
    type: object
...
info:
  contact: {}
  title: Admin-Monolight
paths:
...
  /v1/language:
    post:
      consumes:
      - application/json
      description: This Api for creating a new language
      parameters:
      - description: LanguageCreate Model
        in: body
        name: LanguageCreate
        required: true
        schema:
          $ref: '#/definitions/models.LanguageCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.LanguageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Create Language
      tags:
      - languages
    put:
      consumes:
      - application/json
      description: This Api for updating language
      parameters:
      - description: Update LanguageUpdate Model
        in: body
        name: LanguageUpdate
        required: true
        schema:
          $ref: '#/definitions/models.LanguageUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LanguageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Update Language
      tags:
      - languages
  /v1/language/{id}:
    delete:
      consumes:
      - application/json
      description: This Api for deleting language
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Delete Language
      tags:
      - languages
    get:
      consumes:
      - application/json
      description: This Api for get language
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LanguageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Get Language
      tags:
      - languages
  /v1/languages:
    get:
      consumes:
      - application/json
      description: This Api for get all languages
      parameters:
      - description: Page
        in: query
        name: page
        required: true
        type: integer
      - description: Limit
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.LanguageResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: ListLanguages
      tags:
      - languages
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"